Moved Templight implementation files to lib/ and include/
In order to implement unit tests, we'll have to link templight against
the unit test files -- however, since we only create a templight
executable, not a library, this was impossible.

This patch moves all Templight implementation files to lib/ and
include/, and moves all unused files to utils/.

When the target 'templight' is compiled, a new library called
clangTemplight will be created, against which we can link the unittest
files.